@@ -3,14 +3,14 @@
   "description": "React Native Mapview component for iOS + Android",
   "main": "index.js",
   "author": "Leland Richardson <leland.m.richardson@gmail.com>",
-  "version": "0.24.0",
+  "version": "0.25.0",
   "scripts": {
     "start": "node node_modules/react-native/local-cli/cli.js start",
     "run:packager": "./node_modules/react-native/scripts/packager.sh",
     "run:ios": "react-native run-ios --project-path ./example/ios",
     "start:android": "adb shell am start -n com.airbnb.android.react.maps.example/.MainActivity",
     "run:android": "./gradlew installDebug && npm run start:android",
-    "lint": "eslint ./",
+    "lint": "./node_modules/eslint/bin/eslint.js .",
     "build": "npm run build:js && npm run build:android && npm run build:ios",
     "build:js": "exit 0",
     "build:ios": "bundle install --path ./example/ios/bundles && bundle exec pod install --project-directory=./example/ios/",
@@ -39,20 +39,17 @@
   "devDependencies": {
     "@babel/core": "^7.4.3",
     "@babel/runtime": "^7.4.3",
-    "babel-eslint": "^6.1.2",
+    "@react-native-community/eslint-config": "0.0.5",
+    "babel-eslint": "^10.0.1",
     "babel-plugin-module-resolver": "^3.2.0",
-    "babel-preset-airbnb": "^1.1.1",
-    "babel-preset-react-native": "1.9.0",
-    "eslint": "^3.3.1",
-    "eslint-config-airbnb": "^10.0.1",
-    "eslint-plugin-import": "^1.14.0",
-    "eslint-plugin-jsx-a11y": "^2.1.0",
-    "eslint-plugin-prefer-object-spread": "^1.1.0",
-    "eslint-plugin-react": "^6.1.2",
+    "eslint": "^5.16.0",
+    "eslint-plugin-import": "^2.17.2",
+    "eslint-plugin-jsx-a11y": "^6.2.1",
+    "eslint-plugin-prefer-object-spread": "^1.2.1",
+    "eslint-plugin-react": "^7.13.0",
     "gitbook-cli": "^2.3.0",
     "lodash": "^4.17.2",
     "metro-react-native-babel-preset": "^0.53.1",
-    "prop-types": "^15.7.2",
     "react": "16.8.3",
     "react-native": "0.59.3"
   },
