@@ -2,6 +2,7 @@ import PropTypes from 'prop-types';
 import React from 'react';
 import {
   EdgeInsetsPropType,
+  PointPropType,
   Platform,
   Animated as RNAnimated,
   requireNativeComponent,
@@ -19,6 +20,7 @@ import MapCallout from './MapCallout';
 import MapCalloutSubview from './MapCalloutSubview';
 import MapOverlay from './MapOverlay';
 import MapUrlTile from './MapUrlTile';
+import MapWMSTile from './MapWMSTile';
 import MapLocalTile from './MapLocalTile';
 import AnimatedRegion from './AnimatedRegion';
 import {
@@ -38,10 +40,7 @@ export const MAP_TYPES = {
   MUTEDSTANDARD: 'mutedStandard',
 };
 
-const GOOGLE_MAPS_ONLY_TYPES = [
-  MAP_TYPES.TERRAIN,
-  MAP_TYPES.NONE,
-];
+const GOOGLE_MAPS_ONLY_TYPES = [MAP_TYPES.TERRAIN, MAP_TYPES.NONE];
 
 const viewConfig = {
   uiViewClassName: 'AIR<provider>Map',
@@ -49,7 +48,6 @@ const viewConfig = {
     region: true,
   },
 };
-
 
 /**
  * Defines the map camera.
@@ -65,7 +63,6 @@ const CameraShape = PropTypes.shape({
   zoom: PropTypes.number.isRequired,
 });
 
-
 // if ViewPropTypes is not defined fall back to View.propType (to support RN < 0.44)
 const viewPropTypes = ViewPropTypes || View.propTypes;
 
@@ -76,9 +73,7 @@ const propTypes = {
    * Any value other than "google" will default to using
    * MapKit in iOS or GoogleMaps in android as the map provider.
    */
-  provider: PropTypes.oneOf([
-    'google',
-  ]),
+  provider: PropTypes.oneOf(['google']),
 
   /**
    * Used to style and layout the `MapView`.  See `StyleSheet.js` and
@@ -177,7 +172,7 @@ const propTypes = {
    */
   zoomControlEnabled: PropTypes.bool,
 
-    /**
+  /**
    * If `false` the user won't be able to pinch/rotate the map.
    * Default value is `true`.
    *
@@ -503,6 +498,13 @@ const propTypes = {
   kmlSrc: PropTypes.string,
 
   /**
+   * Offset Point x y for compass location.
+   *
+   * @platform ios
+   */
+  compassOffset: PointPropType,
+
+  /**
    * Callback that is called when a level is activated on a indoor building.
    */
   onIndoorLevelActivated: PropTypes.func,
@@ -531,10 +533,22 @@ class MapView extends React.Component {
     return { provider: this.props.provider };
   }
 
-  componentWillUpdate(nextProps) {
+  getSnapshotBeforeUpdate(prevProps) {
+    if (
+      this.state.isReady &&
+      this.props.customMapStyle !== prevProps.customMapStyle
+    ) {
+      this._updateStyle(this.props);
+    }
+    return this.props.region || null;
+  }
+
+  componentDidUpdate(prevProps, prevState, region) {
     const a = this.__lastRegion;
-    const b = nextProps.region;
-    if (!a || !b) return;
+    const b = region;
+    if (!a || !b) {
+      return;
+    }
     if (
       a.latitude !== b.latitude ||
       a.longitude !== b.longitude ||
@@ -548,13 +562,15 @@ class MapView extends React.Component {
   componentDidMount() {
     const { isReady } = this.state;
     if (isReady) {
-      this._updateStyle();
+      this._updateStyle(this.props);
     }
   }
 
-  _updateStyle() {
-    const { customMapStyle } = this.props;
-    this.map.setNativeProps({ customMapStyleString: JSON.stringify(customMapStyle) });
+  _updateStyle(props) {
+    const { customMapStyle } = props;
+    this.map.setNativeProps({
+      customMapStyleString: JSON.stringify(customMapStyle),
+    });
   }
 
   _onMapReady() {
@@ -564,15 +580,19 @@ class MapView extends React.Component {
     } else if (initialRegion) {
       this.map.setNativeProps({ initialRegion });
     }
-    this._updateStyle();
+    this._updateStyle(this.props);
     this.setState({ isReady: true }, () => {
-      if (onMapReady) onMapReady();
+      if (onMapReady) {
+        onMapReady();
+      }
     });
   }
 
   _onLayout(e) {
     const { layout } = e.nativeEvent;
-    if (!layout.width || !layout.height) return;
+    if (!layout.width || !layout.height) {
+      return;
+    }
     if (this.state.isReady && !this.__layoutCalled) {
       const { region, initialRegion } = this.props;
       if (region) {
@@ -623,8 +643,15 @@ class MapView extends React.Component {
   }
 
   animateToNavigation(location, bearing, angle, duration) {
-    console.warn('animateToNavigation() is deprecated, use animateCamera() instead');
-    this._runCommand('animateToNavigation', [location, bearing, angle, duration || 500]);
+    console.warn(
+      'animateToNavigation() is deprecated, use animateCamera() instead'
+    );
+    this._runCommand('animateToNavigation', [
+      location,
+      bearing,
+      angle,
+      duration || 500,
+    ]);
   }
 
   animateToRegion(region, duration) {
@@ -632,17 +659,23 @@ class MapView extends React.Component {
   }
 
   animateToCoordinate(latLng, duration) {
-    console.warn('animateToCoordinate() is deprecated, use animateCamera() instead');
+    console.warn(
+      'animateToCoordinate() is deprecated, use animateCamera() instead'
+    );
     this._runCommand('animateToCoordinate', [latLng, duration || 500]);
   }
 
   animateToBearing(bearing, duration) {
-    console.warn('animateToBearing() is deprecated, use animateCamera() instead');
+    console.warn(
+      'animateToBearing() is deprecated, use animateCamera() instead'
+    );
     this._runCommand('animateToBearing', [bearing, duration || 500]);
   }
 
   animateToViewingAngle(angle, duration) {
-    console.warn('animateToViewingAngle() is deprecated, use animateCamera() instead');
+    console.warn(
+      'animateToViewingAngle() is deprecated, use animateCamera() instead'
+    );
     this._runCommand('animateToViewingAngle', [angle, duration || 500]);
   }
 
@@ -675,7 +708,9 @@ class MapView extends React.Component {
    */
   async getMapBoundaries() {
     if (Platform.OS === 'android') {
-      return await NativeModules.AirMapModule.getMapBoundaries(this._getHandle());
+      return await NativeModules.AirMapModule.getMapBoundaries(
+        this._getHandle()
+      );
     } else if (Platform.OS === 'ios') {
       return await this._runCommand('getMapBoundaries', []);
     }
@@ -708,12 +743,14 @@ class MapView extends React.Component {
     // For the time being we support the legacy API on iOS.
     // This will be removed in a future release and only the
     // new Promise style API shall be supported.
-    if (Platform.OS === 'ios' && (arguments.length === 4)) {
-      console.warn('Old takeSnapshot API has been deprecated; will be removed in the near future'); //eslint-disable-line
-      const width = arguments[0]; // eslint-disable-line
-      const height = arguments[1]; // eslint-disable-line
-      const region = arguments[2]; // eslint-disable-line
-      const callback = arguments[3]; // eslint-disable-line
+    if (Platform.OS === 'ios' && arguments.length === 4) {
+      console.warn(
+        'Old takeSnapshot API has been deprecated; will be removed in the near future'
+      );
+      const width = arguments[0];
+      const height = arguments[1];
+      const region = arguments[2];
+      const callback = arguments[3];
       this._runCommand('takeSnapshot', [
         width || 0,
         height || 0,
@@ -735,10 +772,12 @@ class MapView extends React.Component {
       quality: args.quality || 1.0,
       result: args.result || 'file',
     };
-    if ((config.format !== 'png') &&
-        (config.format !== 'jpg')) throw new Error('Invalid format specified');
-    if ((config.result !== 'file') &&
-        (config.result !== 'base64')) throw new Error('Invalid result specified');
+    if (config.format !== 'png' && config.format !== 'jpg') {
+      throw new Error('Invalid format specified');
+    }
+    if (config.result !== 'file' && config.result !== 'base64') {
+      throw new Error('Invalid result specified');
+    }
 
     // Call native function
     if (Platform.OS === 'android') {
@@ -776,7 +815,10 @@ class MapView extends React.Component {
    */
   pointForCoordinate(coordinate) {
     if (Platform.OS === 'android') {
-      return NativeModules.AirMapModule.pointForCoordinate(this._getHandle(), coordinate);
+      return NativeModules.AirMapModule.pointForCoordinate(
+        this._getHandle(),
+        coordinate
+      );
     } else if (Platform.OS === 'ios') {
       return this._runCommand('pointForCoordinate', [coordinate]);
     }
@@ -794,7 +836,10 @@ class MapView extends React.Component {
    */
   coordinateForPoint(point) {
     if (Platform.OS === 'android') {
-      return NativeModules.AirMapModule.coordinateForPoint(this._getHandle(), point);
+      return NativeModules.AirMapModule.coordinateForPoint(
+        this._getHandle(),
+        point
+      );
     } else if (Platform.OS === 'ios') {
       return this._runCommand('coordinateForPoint', [point]);
     }
@@ -825,12 +870,12 @@ class MapView extends React.Component {
   boundingBoxForRegion(region) {
     return {
       northEast: {
-        latitude: region.latitude + (region.latitudeDelta / 2),
-        longitude: region.longitude + (region.longitudeDelta / 2),
+        latitude: region.latitude + region.latitudeDelta / 2,
+        longitude: region.longitude + region.longitudeDelta / 2,
       },
       southWest: {
-        latitude: region.latitude - (region.latitudeDelta / 2),
-        longitude: region.longitude - (region.longitudeDelta / 2),
+        latitude: region.latitude - region.latitudeDelta / 2,
+        longitude: region.longitude - region.longitudeDelta / 2,
       },
     };
   }
@@ -899,8 +944,11 @@ class MapView extends React.Component {
         onLayout: this._onLayout,
         ...this.props,
       };
-      if (Platform.OS === 'ios' && props.provider === ProviderConstants.PROVIDER_DEFAULT
-        && GOOGLE_MAPS_ONLY_TYPES.includes(props.mapType)) {
+      if (
+        Platform.OS === 'ios' &&
+        props.provider === ProviderConstants.PROVIDER_DEFAULT &&
+        GOOGLE_MAPS_ONLY_TYPES.includes(props.mapType)
+      ) {
         props.mapType = MAP_TYPES.standard;
       }
       props.handlePanDrag = !!props.onPanDrag;
@@ -919,7 +967,9 @@ class MapView extends React.Component {
     if (Platform.OS === 'android' && this.props.liteMode) {
       return (
         <AIRMapLite
-          ref={ref => { this.map = ref; }}
+          ref={ref => {
+            this.map = ref;
+          }}
           {...props}
         />
       );
@@ -929,7 +979,9 @@ class MapView extends React.Component {
 
     return (
       <AIRMap
-        ref={ref => { this.map = ref; }}
+        ref={ref => {
+          this.map = ref;
+        }}
         {...props}
       />
     );
@@ -942,29 +994,34 @@ MapView.childContextTypes = childContextTypes;
 
 MapView.MAP_TYPES = MAP_TYPES;
 
-const nativeComponent = Component => requireNativeComponent(Component, MapView, {
-  nativeOnly: {
-    onChange: true,
-    onMapReady: true,
-    onKmlReady: true,
-    handlePanDrag: true,
-  },
-});
+const nativeComponent = Component =>
+  requireNativeComponent(Component, MapView, {
+    nativeOnly: {
+      onChange: true,
+      onMapReady: true,
+      onKmlReady: true,
+      handlePanDrag: true,
+    },
+  });
 const airMaps = {
   default: nativeComponent('AIRMap'),
 };
 if (Platform.OS === 'android') {
   airMaps.google = airMaps.default;
 } else {
-  airMaps.google = googleMapIsInstalled ? nativeComponent('AIRGoogleMap') :
-    createNotSupportedComponent('react-native-maps: AirGoogleMaps dir must be added to your xCode project to support GoogleMaps on iOS.'); // eslint-disable-line max-len
+  airMaps.google = googleMapIsInstalled
+    ? nativeComponent('AIRGoogleMap')
+    : createNotSupportedComponent(
+        'react-native-maps: AirGoogleMaps dir must be added to your xCode project to support GoogleMaps on iOS.'
+      );
 }
 const getAirMapComponent = provider => airMaps[provider || 'default'];
 
 let AIRMapLite;
 if (!NativeModules.UIManager.getViewManagerConfig) {
   // RN < 0.58
-  AIRMapLite = NativeModules.UIManager.AIRMapLite &&
+  AIRMapLite =
+    NativeModules.UIManager.AIRMapLite &&
     requireNativeComponent('AIRMapLite', MapView, {
       nativeOnly: {
         onChange: true,
@@ -974,7 +1031,8 @@ if (!NativeModules.UIManager.getViewManagerConfig) {
     });
 } else {
   // RN >= 0.58
-  AIRMapLite = NativeModules.UIManager.getViewManagerConfig('AIRMapLite') &&
+  AIRMapLite =
+    NativeModules.UIManager.getViewManagerConfig('AIRMapLite') &&
     requireNativeComponent('AIRMapLite', MapView, {
       nativeOnly: {
         onChange: true,
@@ -986,7 +1044,9 @@ if (!NativeModules.UIManager.getViewManagerConfig) {
 
 export const Animated = RNAnimated.createAnimatedComponent(MapView);
 
-export const ProviderPropType = PropTypes.oneOf(Object.values(ProviderConstants));
+export const ProviderPropType = PropTypes.oneOf(
+  Object.values(ProviderConstants)
+);
 
 /**
  * TODO:
@@ -1001,6 +1061,7 @@ MapView.Polyline = MapPolyline;
 MapView.Polygon = MapPolygon;
 MapView.Circle = MapCircle;
 MapView.UrlTile = MapUrlTile;
+MapView.MapWMSTile = MapWMSTile;
 MapView.LocalTile = MapLocalTile;
 MapView.Overlay = MapOverlay;
 MapView.Callout = MapCallout;
