@@ -8,7 +8,11 @@ import {
   Platform,
 } from 'react-native';
 
-import MapView, { ProviderPropType, Marker, AnimatedRegion } from 'react-native-maps';
+import MapView, {
+  ProviderPropType,
+  Marker,
+  AnimatedRegion,
+} from 'react-native-maps';
 
 const screen = Dimensions.get('window');
 
@@ -33,8 +37,8 @@ class AnimatedMarkers extends React.Component {
   animate() {
     const { coordinate } = this.state;
     const newCoordinate = {
-      latitude: LATITUDE + ((Math.random() - 0.5) * (LATITUDE_DELTA / 2)),
-      longitude: LONGITUDE + ((Math.random() - 0.5) * (LONGITUDE_DELTA / 2)),
+      latitude: LATITUDE + (Math.random() - 0.5) * (LATITUDE_DELTA / 2),
+      longitude: LONGITUDE + (Math.random() - 0.5) * (LONGITUDE_DELTA / 2),
     };
 
     if (Platform.OS === 'android') {
@@ -60,7 +64,9 @@ class AnimatedMarkers extends React.Component {
           }}
         >
           <Marker.Animated
-            ref={marker => { this.marker = marker; }}
+            ref={marker => {
+              this.marker = marker;
+            }}
             coordinate={this.state.coordinate}
           />
         </MapView>
