@@ -1,11 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 
-import {
-  StyleSheet,
-  Text,
-  Animated,
-} from 'react-native';
+import { StyleSheet, Text, Animated } from 'react-native';
 
 class AnimatedPriceMarker extends React.Component {
   render() {
@@ -38,9 +34,7 @@ class AnimatedPriceMarker extends React.Component {
         <Animated.View
           style={[styles.arrowBorder, { borderTopColor: border }]}
         />
-        <Animated.View
-          style={[styles.arrow, { borderTopColor: background }]}
-        />
+        <Animated.View style={[styles.arrow, { borderTopColor: background }]} />
       </Animated.View>
     );
   }
@@ -98,7 +92,6 @@ const styles = StyleSheet.create({
   },
   selectedArrow: {
     borderTopColor: '#4da2ab',
-
   },
   selectedArrowBorder: {
     borderTopColor: '#007a87',
