@@ -1,29 +1,17 @@
 import React from 'react';
-import {
-  StyleSheet,
-  View,
-  TouchableOpacity,
-  Text,
-  Alert,
-} from 'react-native';
+import { StyleSheet, View, TouchableOpacity, Text, Alert } from 'react-native';
 
 import MapView, { ProviderPropType } from 'react-native-maps';
 
 const LATITUDE = 37.78825;
 const LONGITUDE = -122.4324;
 
-
 class CameraControl extends React.Component {
   async getCamera() {
     const camera = await this.map.getCamera();
-    Alert.alert(
-      'Current camera',
-      JSON.stringify(camera),
-      [
-        { text: 'OK' },
-      ],
-      { cancelable: true }
-    );
+    Alert.alert('Current camera', JSON.stringify(camera), [{ text: 'OK' }], {
+      cancelable: true,
+    });
   }
 
   async setCamera() {
@@ -93,7 +81,6 @@ class CameraControl extends React.Component {
 CameraControl.propTypes = {
   provider: ProviderPropType,
 };
-
 
 const styles = StyleSheet.create({
   container: {
