@@ -1,10 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 
-import {
-  StyleSheet,
-  View,
-} from 'react-native';
+import { StyleSheet, View } from 'react-native';
 
 const propTypes = {
   children: PropTypes.node.isRequired,
@@ -16,9 +13,7 @@ class CustomCallout extends React.Component {
     return (
       <View style={[styles.container, this.props.style]}>
         <View style={styles.bubble}>
-          <View style={styles.amount}>
-            {this.props.children}
-          </View>
+          <View style={styles.amount}>{this.props.children}</View>
         </View>
         <View style={styles.arrowBorder} />
         <View style={styles.arrow} />
