@@ -1,22 +1,15 @@
-/* eslint-disable */
-
 import React from 'react';
 import PropTypes from 'prop-types';
 
-import {
-  View,
-  Animated,
-  PanResponder,
-} from 'react-native';
+import { View, Animated, PanResponder } from 'react-native';
 
 const ModePropType = PropTypes.oneOf(['decay', 'snap', 'spring-origin']);
 const OvershootPropType = PropTypes.oneOf(['spring', 'clamp']);
 const AnimatedPropType = PropTypes.any;
 
-class PanController extends React.Component{
-
+class PanController extends React.Component {
   static propTypes = {
-      // Component Config
+    // Component Config
     lockDirection: PropTypes.bool,
     horizontal: PropTypes.bool,
     vertical: PropTypes.bool,
@@ -29,54 +22,55 @@ class PanController extends React.Component{
     snapSpacingX: PropTypes.number, // TODO: also allow an array of values?
     snapSpacingY: PropTypes.number,
 
-      // Animated Values
+    // Animated Values
     panX: AnimatedPropType,
     panY: AnimatedPropType,
 
-      // Animation Config
+    // Animation Config
     overshootSpringConfig: PropTypes.any,
     momentumDecayConfig: PropTypes.any,
     springOriginConfig: PropTypes.any,
     directionLockDistance: PropTypes.number,
     overshootReductionFactor: PropTypes.number,
 
-      // Events
+    // Events
     onOvershoot: PropTypes.func,
     onDirectionChange: PropTypes.func,
     onReleaseX: PropTypes.func,
     onReleaseY: PropTypes.func,
     onRelease: PropTypes.func,
 
-      //...PanResponderPropTypes,
-  }
+    //...PanResponderPropTypes,
+  };
 
-    // getInitialState() {
-    //     //TODO:
-    //     // it's possible we want to move some props over to state.
-    //     // For example, xBounds/yBounds might need to be
-    //     // calculated/updated automatically
-    //     //
-    //     // This could also be done with a higher-order component
-    //     // that just massages props passed in...
-    //     return {
-    //
-    //     };
-    // },
+  // getInitialState() {
+  //     //TODO:
+  //     // it's possible we want to move some props over to state.
+  //     // For example, xBounds/yBounds might need to be
+  //     // calculated/updated automatically
+  //     //
+  //     // This could also be done with a higher-order component
+  //     // that just massages props passed in...
+  //     return {
+  //
+  //     };
+  // },
 
-  _responder = null
-  _listener = null
-  _direction = null
+  _responder = null;
+  _listener = null;
+  _direction = null;
 
   constructor(props) {
-    super(props)
+    super(props);
 
     this.deceleration = 0.997;
-    if (props.momentumDecayConfig && this.props.momentumDecayConfig.deceleration) {
+    if (
+      props.momentumDecayConfig &&
+      this.props.momentumDecayConfig.deceleration
+    ) {
       this.deceleration = this.props.momentumDecayConfig.deceleration;
     }
   }
-
-
 
   componentWillMount() {
     this._responder = PanResponder.create({
@@ -91,22 +85,23 @@ class PanController extends React.Component{
         this.handleResponderGrant(panX, xMode);
         this.handleResponderGrant(panY, yMode);
 
-        this._direction = horizontal && !vertical ? 'x' : (vertical && !horizontal ? 'y' : null);
+        this._direction =
+          horizontal && !vertical ? 'x' : vertical && !horizontal ? 'y' : null;
       },
 
       onPanResponderMove: (_, { dx, dy, x0, y0 }) => {
         let {
-            panX,
-            panY,
-            xBounds,
-            yBounds,
-            overshootX,
-            overshootY,
-            horizontal,
-            vertical,
-            lockDirection,
-            directionLockDistance,
-            } = this.props;
+          panX,
+          panY,
+          xBounds,
+          yBounds,
+          overshootX,
+          overshootY,
+          horizontal,
+          vertical,
+          lockDirection,
+          directionLockDistance,
+        } = this.props;
 
         if (!this._direction) {
           const dx2 = dx * dx;
@@ -140,46 +135,65 @@ class PanController extends React.Component{
 
       onPanResponderRelease: (_, { vx, vy, dx, dy }) => {
         let {
-            panX,
-            panY,
-            xBounds,
-            yBounds,
-            overshootX,
-            overshootY,
-            horizontal,
-            vertical,
-            lockDirection,
-            xMode,
-            yMode,
-            snapSpacingX,
-            snapSpacingY,
-            } = this.props;
+          panX,
+          panY,
+          xBounds,
+          yBounds,
+          overshootX,
+          overshootY,
+          horizontal,
+          vertical,
+          lockDirection,
+          xMode,
+          yMode,
+          snapSpacingX,
+          snapSpacingY,
+        } = this.props;
 
         let cancel = false;
 
         const dir = this._direction;
 
         if (this.props.onRelease) {
-          cancel = false === this.props.onRelease({ vx, vy, dx, dy });
+          cancel = this.props.onRelease({ vx, vy, dx, dy }) === false;
         }
 
         if (!cancel && horizontal && (!lockDirection || dir === 'x')) {
           let [xMin, xMax] = xBounds;
           if (this.props.onReleaseX) {
-            cancel = false === this.props.onReleaseX({ vx, vy, dx, dy });
+            cancel = this.props.onReleaseX({ vx, vy, dx, dy }) === false;
           }
-          !cancel && this.handleResponderRelease(panX, xMin, xMax, vx, overshootX, xMode, snapSpacingX);
+          !cancel &&
+            this.handleResponderRelease(
+              panX,
+              xMin,
+              xMax,
+              vx,
+              overshootX,
+              xMode,
+              snapSpacingX
+            );
         }
 
         if (!cancel && vertical && (!lockDirection || dir === 'y')) {
           let [yMin, yMax] = yBounds;
           if (this.props.onReleaseY) {
-            cancel = false === this.props.onReleaseY({ vx, vy, dx, dy });
+            cancel = this.props.onReleaseY({ vx, vy, dx, dy }) === false;
           }
-          !cancel && this.handleResponderRelease(panY, yMin, yMax, vy, overshootY, yMode, snapSpacingY);
+          !cancel &&
+            this.handleResponderRelease(
+              panY,
+              yMin,
+              yMax,
+              vy,
+              overshootY,
+              yMode,
+              snapSpacingY
+            );
         }
 
-        this._direction = horizontal && !vertical ? 'x' : (vertical && !horizontal ? 'y' : null);
+        this._direction =
+          horizontal && !vertical ? 'x' : vertical && !horizontal ? 'y' : null;
       },
     });
   }
@@ -211,9 +225,16 @@ class PanController extends React.Component{
     anim.setValue(val);
   }
 
-  handleResponderRelease(anim, min, max, velocity, overshoot, mode, snapSpacing) {
+  handleResponderRelease(
+    anim,
+    min,
+    max,
+    velocity,
+    overshoot,
+    mode,
+    snapSpacing
+  ) {
     anim.flattenOffset();
-
 
     if (anim._value < min) {
       if (this.props.onOvershoot) {
@@ -250,7 +271,14 @@ class PanController extends React.Component{
     } else {
       switch (mode) {
         case 'snap':
-          this.handleSnappedScroll(anim, min, max, velocity, snapSpacing, overshoot);
+          this.handleSnappedScroll(
+            anim,
+            min,
+            max,
+            velocity,
+            snapSpacing,
+            overshoot
+          );
           break;
 
         case 'decay':
@@ -353,7 +381,7 @@ class PanController extends React.Component{
   }
 
   closestCenter(x, spacing) {
-    const plus = (x % spacing) < spacing / 2 ? 0 : spacing;
+    const plus = x % spacing < spacing / 2 ? 0 : spacing;
     return Math.round(x / spacing) * spacing + plus;
   }
 
@@ -364,8 +392,10 @@ class PanController extends React.Component{
 
     while (true) {
       t += 16;
-      x = x0 + (vx / (1 - this.deceleration)) *
-      (1 - Math.exp(-(1 - this.deceleration) * t));
+      x =
+        x0 +
+        (vx / (1 - this.deceleration)) *
+          (1 - Math.exp(-(1 - this.deceleration) * t));
       if (Math.abs(x - x1) < 0.1) {
         x1 = x;
         break;
@@ -382,8 +412,10 @@ class PanController extends React.Component{
     let vf;
     while (true) {
       t += 16;
-      x = x0 + (vx / (1 - this.deceleration)) *
-      (1 - Math.exp(-(1 - this.deceleration) * t));
+      x =
+        x0 +
+        (vx / (1 - this.deceleration)) *
+          (1 - Math.exp(-(1 - this.deceleration) * t));
       vf = (x - x1) / 16;
       if (x > bounds[0] && x < bounds[1]) {
         break;
@@ -396,21 +428,21 @@ class PanController extends React.Component{
     return vf;
   }
 
-// componentDidMount() {
-//    //TODO: we may need to measure the children width/height here?
-// },
-//
-// componentWillUnmount() {
-//
-// },
-//
-// componentDidUnmount() {
-//
-// },
+  // componentDidMount() {
+  //    //TODO: we may need to measure the children width/height here?
+  // },
+  //
+  // componentWillUnmount() {
+  //
+  // },
+  //
+  // componentDidUnmount() {
+  //
+  // },
 
   render() {
     return <View {...this.props} {...this._responder.panHandlers} />;
   }
-};
+}
 
 export default PanController;
