@@ -55,10 +55,7 @@ class PolylineCreator extends React.Component {
       this.setState({
         editing: {
           ...editing,
-          coordinates: [
-            ...editing.coordinates,
-            e.nativeEvent.coordinate,
-          ],
+          coordinates: [...editing.coordinates, e.nativeEvent.coordinate],
         },
       });
     }
@@ -83,7 +80,7 @@ class PolylineCreator extends React.Component {
               strokeWidth={1}
             />
           ))}
-          {this.state.editing &&
+          {this.state.editing && (
             <Polyline
               key="editingPolyline"
               coordinates={this.state.editing.coordinates}
@@ -91,7 +88,7 @@ class PolylineCreator extends React.Component {
               fillColor="rgba(255,0,0,0.5)"
               strokeWidth={1}
             />
-          }
+          )}
         </MapView>
         <View style={styles.buttonContainer}>
           {this.state.editing && (
