@@ -1,10 +1,5 @@
 import React from 'react';
-import {
-  Dimensions,
-  StyleSheet,
-  Text,
-  View,
-} from 'react-native';
+import { Dimensions, StyleSheet, Text, View } from 'react-native';
 
 import MapView, { Marker, ProviderPropType } from 'react-native-maps';
 
@@ -27,8 +22,9 @@ class ZIndexMarkers extends React.Component {
     const markerInfo = [];
     for (let i = 1; i < NUM_MARKERS; i++) {
       markerInfo.push({
-        latitude: (((Math.random() * 2) - 1) * MARKERS_LATITUDE_DELTA) + LATITUDE,
-        longitude: (((Math.random() * 2) - 1) * MARKERS_LONGITUDE_DELTA) + LONGITUDE,
+        latitude: (Math.random() * 2 - 1) * MARKERS_LATITUDE_DELTA + LATITUDE,
+        longitude:
+          (Math.random() * 2 - 1) * MARKERS_LONGITUDE_DELTA + LONGITUDE,
         isSpecial: Math.random() < PERCENT_SPECIAL_MARKERS,
         id: i,
       });
@@ -40,20 +36,22 @@ class ZIndexMarkers extends React.Component {
   }
 
   render() {
-    const markers = this.state.markerInfo.map((markerInfo) =>
+    const markers = this.state.markerInfo.map(markerInfo => (
       <Marker
         coordinate={markerInfo}
         key={markerInfo.id}
         pinColor={markerInfo.isSpecial ? '#c5a620' : null}
         style={markerInfo.isSpecial ? styles.specialMarker : null}
       />
-    );
+    ));
 
     return (
       <View style={styles.container}>
         <MapView
           provider={this.props.provider}
-          ref={ref => { this.map = ref; }}
+          ref={ref => {
+            this.map = ref;
+          }}
           style={styles.map}
           initialRegion={{
             latitude: LATITUDE,
@@ -65,7 +63,10 @@ class ZIndexMarkers extends React.Component {
           {markers}
         </MapView>
         <View style={styles.textContainer}>
-          <Text>The yellow markers have a higher zIndex and appear above other markers.</Text>
+          <Text>
+            The yellow markers have a higher zIndex and appear above other
+            markers.
+          </Text>
         </View>
       </View>
     );
