@@ -1,10 +1,6 @@
 import PropTypes from 'prop-types';
 import React from 'react';
-import {
-  StyleSheet,
-  ViewPropTypes,
-  View,
-} from 'react-native';
+import { StyleSheet, ViewPropTypes, View } from 'react-native';
 import decorateMapComponent, {
   SUPPORTED,
   NOT_SUPPORTED,
@@ -18,16 +14,17 @@ const propTypes = {
   onPress: PropTypes.func,
 };
 
-const defaultProps = {
-};
+const defaultProps = {};
 
 class MapCalloutSubview extends React.Component {
   render() {
     const AIRMapCalloutSubview = this.getAirComponent();
-    return (<AIRMapCalloutSubview
-      {...this.props}
-      style={[styles.calloutSubview, this.props.style]}
-    />);
+    return (
+      <AIRMapCalloutSubview
+        {...this.props}
+        style={[styles.calloutSubview, this.props.style]}
+      />
+    );
   }
 }
 
@@ -35,8 +32,7 @@ MapCalloutSubview.propTypes = propTypes;
 MapCalloutSubview.defaultProps = defaultProps;
 
 const styles = StyleSheet.create({
-  calloutSubview: {
-  },
+  calloutSubview: {},
 });
 
 export default decorateMapComponent(MapCalloutSubview, {
