@@ -1,10 +1,6 @@
 import PropTypes from 'prop-types';
 import React from 'react';
-import {
-  ColorPropType,
-  ViewPropTypes,
-  View,
-} from 'react-native';
+import { ColorPropType, ViewPropTypes, View } from 'react-native';
 import decorateMapComponent, {
   USES_DEFAULT_IMPLEMENTATION,
   SUPPORTED,
@@ -67,11 +63,7 @@ const propTypes = {
    *
    * @platform ios
    */
-  lineCap: PropTypes.oneOf([
-    'butt',
-    'round',
-    'square',
-  ]),
+  lineCap: PropTypes.oneOf(['butt', 'round', 'square']),
 
   /**
    * The line join style to apply to corners of the path.
@@ -79,11 +71,7 @@ const propTypes = {
    *
    * @platform ios
    */
-  lineJoin: PropTypes.oneOf([
-    'miter',
-    'round',
-    'bevel',
-  ]),
+  lineJoin: PropTypes.oneOf(['miter', 'round', 'bevel']),
 
   /**
    * The limiting value that helps avoid spikes at junctions between connected line segments.
@@ -138,7 +126,12 @@ class MapCircle extends React.Component {
   render() {
     const AIRMapCircle = this.getAirComponent();
     return (
-      <AIRMapCircle {...this.props} ref={ref => { this.circle = ref; }} />
+      <AIRMapCircle
+        {...this.props}
+        ref={ref => {
+          this.circle = ref;
+        }}
+      />
     );
   }
 }
