@@ -49,6 +49,11 @@ const propTypes = {
   description: PropTypes.string,
 
   /**
+   * Test ID for end to end test automation
+   */
+  testID: PropTypes.string,
+
+  /**
    * A custom image to be used as the marker's icon. Only local image resources are allowed to be
    * used.
    */
@@ -272,7 +277,10 @@ class MapMarker extends React.Component {
   }
 
   animateMarkerToCoordinate(coordinate, duration) {
-    this._runCommand('animateMarkerToCoordinate', [coordinate, duration || 500]);
+    this._runCommand('animateMarkerToCoordinate', [
+      coordinate,
+      duration || 500,
+    ]);
   }
 
   redraw() {
