@@ -1,11 +1,6 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
-import {
-  View,
-  StyleSheet,
-  Image,
-  Animated,
-} from 'react-native';
+import { View, StyleSheet, Image, Animated } from 'react-native';
 
 import decorateMapComponent, {
   SUPPORTED,
@@ -28,12 +23,13 @@ const propTypes = {
 };
 
 class MapOverlay extends Component {
-
   render() {
     let image;
     if (this.props.image) {
-      if (typeof this.props.image.startsWith === 'function'
-        && this.props.image.startsWith('http')) {
+      if (
+        typeof this.props.image.startsWith === 'function' &&
+        this.props.image.startsWith('http')
+      ) {
         image = this.props.image;
       } else {
         image = Image.resolveAssetSource(this.props.image) || {};
