@@ -1,10 +1,6 @@
 import PropTypes from 'prop-types';
 import React from 'react';
-import {
-  ColorPropType,
-  ViewPropTypes,
-  View,
-} from 'react-native';
+import { ColorPropType, ViewPropTypes, View } from 'react-native';
 import decorateMapComponent, {
   USES_DEFAULT_IMPLEMENTATION,
   SUPPORTED,
@@ -19,24 +15,30 @@ const propTypes = {
   /**
    * An array of coordinates to describe the polygon
    */
-  coordinates: PropTypes.arrayOf(PropTypes.shape({
-    /**
-     * Latitude/Longitude coordinates
-     */
-    latitude: PropTypes.number.isRequired,
-    longitude: PropTypes.number.isRequired,
-  })),
+  coordinates: PropTypes.arrayOf(
+    PropTypes.shape({
+      /**
+       * Latitude/Longitude coordinates
+       */
+      latitude: PropTypes.number.isRequired,
+      longitude: PropTypes.number.isRequired,
+    })
+  ),
 
   /**
    * An array of array of coordinates to describe the polygon holes
    */
-  holes: PropTypes.arrayOf(PropTypes.arrayOf(PropTypes.shape({
-    /**
-     * Latitude/Longitude coordinates
-     */
-    latitude: PropTypes.number.isRequired,
-    longitude: PropTypes.number.isRequired,
-  }))),
+  holes: PropTypes.arrayOf(
+    PropTypes.arrayOf(
+      PropTypes.shape({
+        /**
+         * Latitude/Longitude coordinates
+         */
+        latitude: PropTypes.number.isRequired,
+        longitude: PropTypes.number.isRequired,
+      })
+    )
+  ),
 
   /**
    * Callback that is called when the user presses on the polygon
@@ -79,11 +81,7 @@ const propTypes = {
    *
    * @platform ios
    */
-  lineCap: PropTypes.oneOf([
-    'butt',
-    'round',
-    'square',
-  ]),
+  lineCap: PropTypes.oneOf(['butt', 'round', 'square']),
 
   /**
    * The line join style to apply to corners of the path.
@@ -91,11 +89,7 @@ const propTypes = {
    *
    * @platform ios
    */
-  lineJoin: PropTypes.oneOf([
-    'miter',
-    'round',
-    'bevel',
-  ]),
+  lineJoin: PropTypes.oneOf(['miter', 'round', 'bevel']),
 
   /**
    * The limiting value that helps avoid spikes at junctions between connected line segments.
@@ -159,7 +153,12 @@ class MapPolygon extends React.Component {
   render() {
     const AIRMapPolygon = this.getAirComponent();
     return (
-      <AIRMapPolygon {...this.props} ref={ref => { this.polygon = ref; }} />
+      <AIRMapPolygon
+        {...this.props}
+        ref={ref => {
+          this.polygon = ref;
+        }}
+      />
     );
   }
 }
