@@ -1,10 +1,7 @@
 import PropTypes from 'prop-types';
 import React from 'react';
 
-import {
-  ViewPropTypes,
-  View,
-} from 'react-native';
+import { ViewPropTypes, View } from 'react-native';
 
 import decorateMapComponent, {
   USES_DEFAULT_IMPLEMENTATION,
@@ -69,11 +66,7 @@ const propTypes = {
 class MapUrlTile extends React.Component {
   render() {
     const AIRMapUrlTile = this.getAirComponent();
-    return (
-      <AIRMapUrlTile
-        {...this.props}
-      />
-    );
+    return <AIRMapUrlTile {...this.props} />;
   }
 }
 
