@@ -14,6 +14,7 @@
 #import "AIRGoogleMapPolyline.h"
 #import "AIRGoogleMapCircle.h"
 #import "AIRGoogleMapUrlTile.h"
+#import "AIRGoogleMapWMSTile.h"
 #import "AIRGoogleMapOverlay.h"
 #import <GoogleMaps/GoogleMaps.h>
 #import <MapKit/MapKit.h>
@@ -51,6 +52,8 @@ id regionAsJSON(MKCoordinateRegion region) {
 
 - (id)eventFromCoordinate:(CLLocationCoordinate2D)coordinate;
 
+@property (nonatomic, strong) NSMutableDictionary<NSNumber *, NSDictionary*> *origGestureRecognizersMeta;
+
 @end
 
 @implementation AIRGoogleMap
@@ -61,7 +64,6 @@ id regionAsJSON(MKCoordinateRegion region) {
   BOOL _initialCameraSetOnLoad;
   BOOL _didCallOnMapReady;
   BOOL _didMoveToWindow;
-  NSMutableDictionary<NSNumber *, NSDictionary*> *_origGestureRecognizersMeta;
   BOOL _zoomTapEnabled;
 }
 
@@ -90,7 +92,7 @@ id regionAsJSON(MKCoordinateRegion region) {
               options:NSKeyValueObservingOptionNew
               context:NULL];
       
-    _origGestureRecognizersMeta = [[NSMutableDictionary alloc] init];
+    self.origGestureRecognizersMeta = [[NSMutableDictionary alloc] init];
   }
   return self;
 }
@@ -142,6 +144,10 @@ id regionAsJSON(MKCoordinateRegion region) {
     AIRGoogleMapUrlTile *tile = (AIRGoogleMapUrlTile*)subview;
     tile.tileLayer.map = self;
     [self.tiles addObject:tile];
+  } else if ([subview isKindOfClass:[AIRGoogleMapWMSTile class]]) {
+    AIRGoogleMapWMSTile *tile = (AIRGoogleMapWMSTile*)subview;
+    tile.tileLayer.map = self;
+    [self.tiles addObject:tile];
   } else if ([subview isKindOfClass:[AIRGoogleMapOverlay class]]) {
     AIRGoogleMapOverlay *overlay = (AIRGoogleMapOverlay*)subview;
     overlay.overlay.map = self;
@@ -180,6 +186,10 @@ id regionAsJSON(MKCoordinateRegion region) {
     [self.circles removeObject:circle];
   } else if ([subview isKindOfClass:[AIRGoogleMapUrlTile class]]) {
     AIRGoogleMapUrlTile *tile = (AIRGoogleMapUrlTile*)subview;
+    tile.tileLayer.map = nil;
+    [self.tiles removeObject:tile];
+  } else if ([subview isKindOfClass:[AIRGoogleMapWMSTile class]]) {
+    AIRGoogleMapWMSTile *tile = (AIRGoogleMapWMSTile*)subview;
     tile.tileLayer.map = nil;
     [self.tiles removeObject:tile];
   } else if ([subview isKindOfClass:[AIRGoogleMapOverlay class]]) {
@@ -610,6 +620,7 @@ id regionAsJSON(MKCoordinateRegion region) {
             }
         }
     }
+    free(ivars);
     return action;
 }
 
@@ -619,7 +630,7 @@ id regionAsJSON(MKCoordinateRegion region) {
     NSArray* grs = view.gestureRecognizers;
     for (UIGestureRecognizer* gestureRecognizer in grs) {
         NSNumber* grHash = [NSNumber numberWithUnsignedInteger:gestureRecognizer.hash];
-        if([_origGestureRecognizersMeta objectForKey:grHash] != nil)
+        if([self.origGestureRecognizersMeta objectForKey:grHash] != nil)
             continue; //already patched
         
         //get original handlers
@@ -630,7 +641,10 @@ id regionAsJSON(MKCoordinateRegion region) {
             NSObject* target = [trg valueForKey:@"target"];
             SEL action = [self getActionForTarget:trg];
             isZoomTapGesture = [NSStringFromSelector(action) isEqualToString:@"handleZoomTapGesture:"];
-            [origTargetsActions addObject:@{@"target": target, @"action": NSStringFromSelector(action)}];
+            [origTargetsActions addObject:@{
+                                            @"target": [NSValue valueWithNonretainedObject:target],
+                                            @"action": NSStringFromSelector(action)
+                                            }];
         }
         if (isZoomTapGesture && self.zoomTapEnabled == NO) {
             [view removeGestureRecognizer:gestureRecognizer];
@@ -639,15 +653,16 @@ id regionAsJSON(MKCoordinateRegion region) {
         
         //replace with extendedMapGestureHandler
         for (NSDictionary* origTargetAction in origTargetsActions) {
-            NSObject* target = [origTargetAction objectForKey:@"target"];
+            NSValue* targetValue = [origTargetAction objectForKey:@"target"];
+            NSObject* target = [targetValue nonretainedObjectValue];
             NSString* actionString = [origTargetAction objectForKey:@"action"];
             SEL action = NSSelectorFromString(actionString);
             [gestureRecognizer removeTarget:target action:action];
         }
         [gestureRecognizer addTarget:self action:@selector(extendedMapGestureHandler:)];
         
-        [_origGestureRecognizersMeta setObject:@{@"targets": origTargetsActions}
-                                        forKey:grHash];
+        [self.origGestureRecognizersMeta setObject:@{@"targets": origTargetsActions}
+                                            forKey:grHash];
     }
 }
 
@@ -725,10 +740,11 @@ id regionAsJSON(MKCoordinateRegion region) {
     }
     
     if (performOriginalActions) {
-        NSDictionary* origMeta = [_origGestureRecognizersMeta objectForKey:grHash];
+        NSDictionary* origMeta = [self.origGestureRecognizersMeta objectForKey:grHash];
         NSDictionary* origTargets = [origMeta objectForKey:@"targets"];
         for (NSDictionary* origTarget in origTargets) {
-            NSObject* target = [origTarget objectForKey:@"target"];
+            NSValue* targetValue = [origTarget objectForKey:@"target"];
+            NSObject* target = [targetValue nonretainedObjectValue];
             NSString* actionString = [origTarget objectForKey:@"action"];
             SEL action = NSSelectorFromString(actionString);
 #pragma clang diagnostic push
